requirements.txt
geventhttpclient==1.5.3
keras==2.9.0
matplotlib==3.5.2
numpy==1.22.4
onnx==1.11.0
onnxruntime==1.11.1
pandas==1.4.2
tensorflow==2.9.1
fastapi==0.78.0
uvicorn==0.17.6
fastapi-utils==0.2.1
Pillow==7.0.0
tritonclient==2.22.0
pydantic==1.9.1
tf2onnx==1.10.1
